Add solution for problem `B - New Theatre Square` Link: https://codeforces.com/contest/1359/problem/B Tags: brute force, dp, greedy, implementation, two pointers, *1000 Ref: https://codeforces.com/contest/1359/submission/116283140

+++ codeforces/1359/B.cpp
@@ -120,7 +120,7 @@ void solve()
         int one=0,two=0;
         
         one += count(s.begin(), s.end(), '.');
-        int pos =s.find("..",0);
+        size_t pos =s.find("..",0);
         while (pos != string:: npos)
         {
             two++;
